pyproject.toml
[project]
name = "my-ollama-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.17.2",
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.6",
    "pgvector>=0.4.2",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
]
